Fix AsyncAppender/SocketAppender unit tests
AsyncAppenderBaseTest: The appender was not allowed enough time to
flush the appender on #stop(), causing the empty-check of the appender's
queue to fail. The test appender's max-flush time is now extended from
30ms to >1000ms.

AbstractSocketAppenderIntegrationTest: A unit test expected
deque.takeFirst() to be called only once after a log-event is appended,
but the call actually occurs twice in a while-loop, where the second
call is intended to block, waiting for the next log-event.

AbstractSocketAppenderTest: A unit test that verified appender
behavior when its deque was at capacity neglected to fill the deque
before testing for errors.